+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 All notable changes to `laravel-i18n` will be documented in this file.
 
-## 0.1.6 - 2018-01-28
+## 0.1.7 - 2018-01-30
+* `forceRegionPrefix()` works also on resources
+
+## 0.1.6 - 2018-01-29
 * Improved route registration logic
-* On routes you can call `forceRegionPrefix()` to force region when the `default` setting is set.
+* On routes you can call `forceRegionPrefix()` to force region when the `default` setting is set
 
 ## 0.1.5 - 2018-01-26
 * I18n: manually set the current region with `setRegion($region)`